+++ pallets/circuit/src/tests.rs
@@ -27,10 +27,10 @@ use crate::{mock::*, state::*};
 
 use sp_io::TestExternalities;
 
-use codec::Encode;
+use codec::{Decode, Encode};
 use sp_runtime::AccountId32;
 use sp_std::prelude::*;
-use t3rn_primitives::volatile::LocalState;
+use t3rn_primitives::{volatile::LocalState, xtx::XtxId};
 // use crate::mock;
 
 pub const ALICE: AccountId32 = AccountId32::new([1u8; 32]);
@@ -42,7 +42,7 @@ fn set_ids(
     valid_side_effect: SideEffect<AccountId32, BlockNumber, BalanceOf>,
 ) -> (sp_core::H256, sp_core::H256) {
     let xtx_id: sp_core::H256 =
-        hex!("d02e1b7d4ee308b8b0fad924fd35bb3688760fc986bc1b44e8f123f17aed8a7a").into();
+        hex!("c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6").into();
 
     let side_effect_a_id = valid_side_effect.generate_id::<crate::SystemHashing<Test>>();
 
@@ -125,7 +125,7 @@ fn on_extrinsic_trigger_works_with_single_transfer_not_insured() {
                         phase: Phase::Initialization,
                         event: Event::Circuit(crate::Event::<Test>::XTransactionReadyForExec(
                             hex!(
-                                "d02e1b7d4ee308b8b0fad924fd35bb3688760fc986bc1b44e8f123f17aed8a7a"
+                                "c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6"
                             )
                             .into()
                         )),
@@ -138,7 +138,7 @@ fn on_extrinsic_trigger_works_with_single_transfer_not_insured() {
                                 "0101010101010101010101010101010101010101010101010101010101010101"
                             )),
                             hex!(
-                                "d02e1b7d4ee308b8b0fad924fd35bb3688760fc986bc1b44e8f123f17aed8a7a"
+                                "c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6"
                             )
                             .into(),
                             vec![SideEffect {
@@ -167,7 +167,7 @@ fn on_extrinsic_trigger_works_with_single_transfer_not_insured() {
                 ]
             );
             let xtx_id: sp_core::H256 =
-                hex!("d02e1b7d4ee308b8b0fad924fd35bb3688760fc986bc1b44e8f123f17aed8a7a").into();
+                hex!("c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6").into();
             let side_effect_a_id =
                 valid_transfer_side_effect.generate_id::<crate::SystemHashing<Test>>();
 
@@ -182,7 +182,7 @@ fn on_extrinsic_trigger_works_with_single_transfer_not_insured() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::Ready,
                     total_reward: Some(fee),
@@ -195,6 +195,7 @@ fn on_extrinsic_trigger_works_with_single_transfer_not_insured() {
                 vec![vec![FullSideEffect {
                     input: valid_transfer_side_effect,
                     confirmed: None,
+                    security_lvl: SecurityLvl::Dirty,
                 }]]
             );
         });
@@ -290,7 +291,7 @@ fn on_extrinsic_trigger_emit_works_with_single_transfer_insured() {
                         phase: Phase::Initialization,
                         event: Event::Circuit(crate::Event::<Test>::XTransactionReceivedForExec(
                             hex!(
-                                "d02e1b7d4ee308b8b0fad924fd35bb3688760fc986bc1b44e8f123f17aed8a7a"
+                                "c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6"
                             )
                             .into()
                         )),
@@ -303,7 +304,7 @@ fn on_extrinsic_trigger_emit_works_with_single_transfer_insured() {
                                 "0101010101010101010101010101010101010101010101010101010101010101"
                             )),
                             hex!(
-                                "d02e1b7d4ee308b8b0fad924fd35bb3688760fc986bc1b44e8f123f17aed8a7a"
+                                "c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6"
                             )
                             .into(),
                             vec![SideEffect {
@@ -403,7 +404,7 @@ fn on_extrinsic_trigger_apply_works_with_single_transfer_insured() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::PendingInsurance,
                     total_reward: Some(fee),
@@ -416,6 +417,7 @@ fn on_extrinsic_trigger_apply_works_with_single_transfer_insured() {
                 vec![vec![FullSideEffect {
                     input: valid_transfer_side_effect,
                     confirmed: None,
+                    security_lvl: SecurityLvl::Optimistic,
                 }]]
             );
         });
@@ -487,7 +489,7 @@ fn circuit_handles_insurance_deposit_for_transfers() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::PendingInsurance,
                     total_reward: Some(fee),
@@ -500,6 +502,7 @@ fn circuit_handles_insurance_deposit_for_transfers() {
                 vec![vec![FullSideEffect {
                     input: valid_transfer_side_effect.clone(),
                     confirmed: None,
+                    security_lvl: SecurityLvl::Optimistic,
                 }]]
             );
 
@@ -533,7 +536,7 @@ fn circuit_handles_insurance_deposit_for_transfers() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::Ready,
                     total_reward: Some(fee),
@@ -630,7 +633,7 @@ fn circuit_handles_dirty_swap_with_no_insurance() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::Ready,
                     total_reward: Some(fee),
@@ -643,6 +646,7 @@ fn circuit_handles_dirty_swap_with_no_insurance() {
                 vec![vec![FullSideEffect {
                     input: valid_swap_side_effect.clone(),
                     confirmed: None,
+                    security_lvl: SecurityLvl::Dirty,
                 }]]
             );
 
@@ -752,7 +756,7 @@ fn circuit_handles_swap_with_insurance() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::PendingInsurance,
                     total_reward: Some(fee),
@@ -765,6 +769,7 @@ fn circuit_handles_swap_with_insurance() {
                 vec![vec![FullSideEffect {
                     input: valid_swap_side_effect.clone(),
                     confirmed: None,
+                    security_lvl: SecurityLvl::Optimistic,
                 }]]
             );
 
@@ -798,7 +803,7 @@ fn circuit_handles_swap_with_insurance() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::Ready,
                     total_reward: Some(fee),
@@ -1003,7 +1008,7 @@ fn circuit_handles_add_liquidity_with_insurance() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::PendingInsurance,
                     total_reward: Some(fee),
@@ -1016,6 +1021,7 @@ fn circuit_handles_add_liquidity_with_insurance() {
                 vec![vec![FullSideEffect {
                     input: valid_add_liquidity_side_effect.clone(),
                     confirmed: None,
+                    security_lvl: SecurityLvl::Optimistic,
                 }]]
             );
 
@@ -1049,7 +1055,7 @@ fn circuit_handles_add_liquidity_with_insurance() {
                     requester: AccountId32::new(hex!(
                         "0101010101010101010101010101010101010101010101010101010101010101"
                     )),
-                    timeouts_at: None,
+                    timeouts_at: 100u64,
                     delay_steps_at: None,
                     status: CircuitStatus::Ready,
                     total_reward: Some(fee),
@@ -1093,19 +1099,131 @@ fn circuit_handles_add_liquidity_with_insurance() {
         });
 }
 
-// Ignored for now because of a bug described in t3rn#261
+// fn successfully_confirm_optimistic(side_effect: SideEffect<AccountId32, BlockNumber, BalanceOf>) {
+//
+//     let from = side_effect.encoded_args[0].clone();
+//     let to = side_effect.encoded_args[1].clone();
+//     let amount = side_effect.encoded_args[2].clone();
+//
+//     let mut encoded_balance_transfer_event_1 = pallet_balances::Event::<Test>::Transfer {
+//         from: from.into(), // variant A
+//         to: to.into(), // variant B (dest)
+//         amount: amount.into(), // variant A
+//     }
+//         .encode();
+//
+//     // Adding 4 since Balances Pallet = 4 in construct_runtime! enum
+//     let mut encoded_event_1 = vec![4];
+//     encoded_event_1.append(&mut encoded_balance_transfer_event_1);
+//     let confirmation_transfer_1 =
+//         ConfirmedSideEffect::<AccountId32, BlockNumber, BalanceOf> {
+//             err: None,
+//             output: None,
+//             encoded_effect: encoded_event_1,
+//             inclusion_proof: None,
+//             executioner: BOB_RELAYER,
+//             received_at: 0,
+//             cost: None,
+//         };
+//
+//     assert_ok!(Circuit::confirm_side_effect(
+//         origin_relayer_bob.clone(),
+//         xtx_id.clone(),
+//         valid_transfer_side_effect_1,
+//         confirmation_transfer_1,
+//         None,
+//         None,
+//     ));
+//
+// }
+
+fn successfully_confirm_dirty(
+    side_effect: SideEffect<AccountId32, BlockNumber, BalanceOf>,
+    xtx_id: XtxId<Test>,
+    relayer: AccountId32,
+) {
+    let from = side_effect.encoded_args[0].clone();
+    let to = side_effect.encoded_args[1].clone();
+    let amount = side_effect.encoded_args[2].clone();
+
+    let mut encoded_balance_transfer_event = pallet_balances::Event::<Test>::Transfer {
+        from: Decode::decode(&mut &from[..]).unwrap(), // variant A
+        to: Decode::decode(&mut &to[..]).unwrap(),     // variant A
+        amount: Decode::decode(&mut &amount[..]).unwrap(), // variant A
+    }
+    .encode();
+
+    // Adding 4 since Balances Pallet = 4 in construct_runtime! enum
+    let mut encoded_event = vec![4];
+    encoded_event.append(&mut encoded_balance_transfer_event);
+    let confirmation_transfer = ConfirmedSideEffect::<AccountId32, BlockNumber, BalanceOf> {
+        err: None,
+        output: None,
+        encoded_effect: encoded_event,
+        inclusion_proof: None,
+        executioner: BOB_RELAYER,
+        received_at: 0,
+        cost: None,
+    };
+
+    assert_ok!(Circuit::confirm_side_effect(
+        Origin::signed(relayer),
+        xtx_id.clone(),
+        side_effect,
+        confirmation_transfer,
+        None,
+        None,
+    ));
+}
+
+fn successfully_bond_optimistic(
+    side_effect: SideEffect<AccountId32, BlockNumber, BalanceOf>,
+    xtx_id: XtxId<Test>,
+    relayer: AccountId32,
+    submitter: AccountId32,
+) {
+    let optional_insurance = side_effect.encoded_args[3].clone();
+
+    assert!(
+        optional_insurance.len() == 32,
+        "Wrong test value - optimistic transfer assumes optimistic arguments"
+    );
+
+    assert_ok!(Circuit::bond_insurance_deposit(
+        Origin::signed(relayer.clone()),
+        xtx_id,
+        side_effect.generate_id::<crate::SystemHashing<Test>>(),
+    ));
+
+    let [insurance, reward]: [u128; 2] = Decode::decode(&mut &optional_insurance[..]).unwrap();
+
+    assert_eq!(
+        Circuit::get_insurance_deposits(
+            xtx_id,
+            side_effect.generate_id::<crate::SystemHashing<Test>>()
+        )
+        .unwrap(),
+        InsuranceDeposit {
+            insurance: insurance as u64,
+            reward: reward as u64,
+            requester: Decode::decode(&mut &submitter.encode()[..]).unwrap(),
+            bonded_relayer: Some(relayer.clone()),
+            status: CircuitStatus::Bonded,
+            requested_at: 1,
+        }
+    );
+}
+
 #[test]
-#[ignore]
-fn circuit_handles_transfer_and_swap() {
+fn two_dirty_and_three_optimistic_transfers_are_allocated_to_3_steps_and_all_5_is_confirmed() {
     let origin = Origin::signed(ALICE); // Only sudo access to register new gateways for now
 
-    let origin_relayer_bob = Origin::signed(BOB_RELAYER); // Only sudo access to register new gateways for now
+    let _origin_relayer_bob = Origin::signed(BOB_RELAYER); // Only sudo access to register new gateways for now
 
     let transfer_protocol_box = ExtBuilder::get_transfer_protocol_box();
-    let swap_protocol_box = ExtBuilder::get_swap_protocol_box();
 
     let mut local_state = LocalState::new();
-    let valid_transfer_side_effect = produce_and_validate_side_effect(
+    let valid_transfer_side_effect_1 = produce_and_validate_side_effect(
         vec![
             (Type::Address(32), ArgVariant::A),
             (Type::Address(32), ArgVariant::B),
@@ -1113,26 +1231,225 @@ fn circuit_handles_transfer_and_swap() {
             (Type::Bytes(0), ArgVariant::A), // empty bytes instead of insurance
         ],
         &mut local_state,
-        transfer_protocol_box,
+        transfer_protocol_box.clone(),
     );
 
-    let valid_swap_side_effect = produce_and_validate_side_effect(
+    let valid_transfer_side_effect_2 = produce_and_validate_side_effect(
         vec![
-            (Type::Address(32), ArgVariant::A), // caller
-            (Type::Address(32), ArgVariant::B), // to
-            (Type::Uint(64), ArgVariant::A),    // amount_from
-            (Type::Uint(64), ArgVariant::B),    // amount_to
-            (Type::Bytes(4), ArgVariant::A),    // asset_from
-            (Type::Bytes(4), ArgVariant::B),    // asset_to
-            (Type::Bytes(0), ArgVariant::A),    // no insurance
+            (Type::Address(32), ArgVariant::B),
+            (Type::Address(32), ArgVariant::A),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::Bytes(0), ArgVariant::A), // empty bytes instead of insurance
         ],
         &mut local_state,
-        swap_protocol_box,
+        transfer_protocol_box.clone(),
+    );
+
+    let valid_optimistic_transfer_side_effect_3 = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::C),
+            (Type::Address(32), ArgVariant::A),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::OptionalInsurance, ArgVariant::A), // empty bytes instead of insurance
+        ],
+        &mut local_state,
+        transfer_protocol_box.clone(),
+    );
+
+    let valid_optimistic_transfer_side_effect_4 = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::C),
+            (Type::Address(32), ArgVariant::B),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::OptionalInsurance, ArgVariant::B), // empty bytes instead of insurance
+        ],
+        &mut local_state,
+        transfer_protocol_box.clone(),
+    );
+
+    let valid_optimistic_transfer_side_effect_5 = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::C),
+            (Type::Address(32), ArgVariant::B),
+            (Type::Uint(64), ArgVariant::B),
+            (Type::OptionalInsurance, ArgVariant::B), // empty bytes instead of insurance
+        ],
+        &mut local_state,
+        transfer_protocol_box.clone(),
     );
 
     let side_effects = vec![
-        valid_transfer_side_effect.clone(),
-        valid_swap_side_effect.clone(),
+        valid_transfer_side_effect_1.clone(),
+        valid_transfer_side_effect_2.clone(),
+        valid_optimistic_transfer_side_effect_3.clone(),
+        valid_optimistic_transfer_side_effect_4.clone(),
+        valid_optimistic_transfer_side_effect_5.clone(),
+    ];
+
+    let fee = 1;
+    let sequential = true;
+
+    ExtBuilder::default()
+        .with_standard_side_effects()
+        .with_default_xdns_records()
+        .build()
+        .execute_with(|| {
+            let _ = Balances::deposit_creating(&ALICE, 50);
+            let _ = Balances::deposit_creating(&BOB_RELAYER, 50);
+
+            System::set_block_number(1);
+
+            assert_ok!(Circuit::on_extrinsic_trigger(
+                origin,
+                side_effects,
+                fee,
+                sequential,
+            ));
+
+            let xtx_id: sp_core::H256 =
+                hex!("c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6").into();
+
+            // Confirmation start - 3
+            successfully_bond_optimistic(
+                valid_optimistic_transfer_side_effect_3.clone(),
+                xtx_id.clone(),
+                BOB_RELAYER,
+                ALICE,
+            );
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::PendingInsurance
+            );
+
+            // Confirmation start - 4
+            successfully_bond_optimistic(
+                valid_optimistic_transfer_side_effect_4.clone(),
+                xtx_id.clone(),
+                BOB_RELAYER,
+                ALICE,
+            );
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::PendingInsurance
+            );
+
+            // Confirmation start - 5
+            successfully_bond_optimistic(
+                valid_optimistic_transfer_side_effect_5.clone(),
+                xtx_id.clone(),
+                BOB_RELAYER,
+                ALICE,
+            );
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::Ready
+            );
+
+            // Confirmation start - 3
+            successfully_confirm_dirty(
+                valid_optimistic_transfer_side_effect_3,
+                xtx_id.clone(),
+                BOB_RELAYER,
+            );
+
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::Bonded
+            );
+
+            // Confirmation start - 4
+            successfully_confirm_dirty(
+                valid_optimistic_transfer_side_effect_4,
+                xtx_id.clone(),
+                BOB_RELAYER,
+            );
+
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::Bonded
+            );
+
+            // Confirmation start - 5
+            successfully_confirm_dirty(
+                valid_optimistic_transfer_side_effect_5,
+                xtx_id.clone(),
+                BOB_RELAYER,
+            );
+
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::Finished
+            );
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone())
+                    .unwrap()
+                    .steps_cnt,
+                (1, 3)
+            );
+            // Confirmation start - 1
+            successfully_confirm_dirty(valid_transfer_side_effect_1, xtx_id.clone(), BOB_RELAYER);
+
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::Finished
+            );
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone())
+                    .unwrap()
+                    .steps_cnt,
+                (2, 3)
+            );
+
+            // Confirmation start - 2
+            successfully_confirm_dirty(valid_transfer_side_effect_2, xtx_id.clone(), BOB_RELAYER);
+
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone()).unwrap().status,
+                CircuitStatus::FinishedAllSteps
+            );
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id.clone())
+                    .unwrap()
+                    .steps_cnt,
+                (3, 3)
+            );
+        });
+}
+
+#[test]
+fn two_dirty_transfers_are_allocated_to_2_steps_and_can_be_confirmed() {
+    let origin = Origin::signed(ALICE); // Only sudo access to register new gateways for now
+
+    let _origin_relayer_bob = Origin::signed(BOB_RELAYER); // Only sudo access to register new gateways for now
+
+    let transfer_protocol_box = ExtBuilder::get_transfer_protocol_box();
+
+    let mut local_state = LocalState::new();
+    let valid_transfer_side_effect_1 = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::A),
+            (Type::Address(32), ArgVariant::B),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::Bytes(0), ArgVariant::A), // empty bytes instead of insurance
+        ],
+        &mut local_state,
+        transfer_protocol_box.clone(),
+    );
+
+    let valid_transfer_side_effect_2 = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::B),
+            (Type::Address(32), ArgVariant::A),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::Bytes(0), ArgVariant::A), // empty bytes instead of insurance
+        ],
+        &mut local_state,
+        transfer_protocol_box,
+    );
+
+    let side_effects = vec![
+        valid_transfer_side_effect_1.clone(),
+        valid_transfer_side_effect_2.clone(),
     ];
     let fee = 1;
     let sequential = true;
@@ -1157,16 +1474,106 @@ fn circuit_handles_transfer_and_swap() {
             assert_eq!(events.len(), 8);
 
             let xtx_id: sp_core::H256 =
-                hex!("d02e1b7d4ee308b8b0fad924fd35bb3688760fc986bc1b44e8f123f17aed8a7a").into();
+                hex!("c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6").into();
+
+            // Confirmation start - 1
+            successfully_confirm_dirty(valid_transfer_side_effect_1, xtx_id.clone(), BOB_RELAYER);
+
+            // Confirmation start - 2
+            successfully_confirm_dirty(valid_transfer_side_effect_2, xtx_id.clone(), BOB_RELAYER);
+        });
+}
+
+// ToDo: Order for multiple should now be fixed - verify t3rn#261 is solved
+#[test]
+#[ignore]
+fn circuit_handles_transfer_dirty_and_optimistic_and_swap() {
+    let origin = Origin::signed(ALICE); // Only sudo access to register new gateways for now
+
+    let origin_relayer_bob = Origin::signed(BOB_RELAYER); // Only sudo access to register new gateways for now
+
+    let transfer_protocol_box = ExtBuilder::get_transfer_protocol_box();
+    let swap_protocol_box = ExtBuilder::get_swap_protocol_box();
+
+    let mut local_state = LocalState::new();
+    let valid_transfer_side_effect_1 = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::A),
+            (Type::Address(32), ArgVariant::B),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::Bytes(0), ArgVariant::A), // empty bytes instead of insurance
+        ],
+        &mut local_state,
+        transfer_protocol_box.clone(),
+    );
+
+    let valid_transfer_side_effect_2 = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::A),
+            (Type::Address(32), ArgVariant::B),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::OptionalInsurance, ArgVariant::A),
+        ],
+        &mut local_state,
+        transfer_protocol_box,
+    );
+
+    let valid_swap_side_effect = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::A), // caller
+            (Type::Address(32), ArgVariant::B), // to
+            (Type::Uint(64), ArgVariant::A),    // amount_from
+            (Type::Uint(64), ArgVariant::B),    // amount_to
+            (Type::Bytes(4), ArgVariant::A),    // asset_from
+            (Type::Bytes(4), ArgVariant::B),    // asset_to
+            (Type::Bytes(0), ArgVariant::A),    // no insurance
+        ],
+        &mut local_state,
+        swap_protocol_box,
+    );
+
+    let side_effects = vec![
+        valid_transfer_side_effect_1.clone(),
+        valid_transfer_side_effect_2.clone(),
+        valid_swap_side_effect.clone(),
+    ];
+    let fee = 1;
+    let sequential = true;
+
+    ExtBuilder::default()
+        .with_standard_side_effects()
+        .with_default_xdns_records()
+        .build()
+        .execute_with(|| {
+            let _ = Balances::deposit_creating(&ALICE, 10);
+
+            System::set_block_number(1);
+
+            assert_ok!(Circuit::on_extrinsic_trigger(
+                origin,
+                side_effects,
+                fee,
+                sequential,
+            ));
+
+            let events = System::events();
+            assert_eq!(events.len(), 9);
+
+            let xtx_id: sp_core::H256 =
+                hex!("c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6").into();
 
             // Confirmation start
-            let encoded_balance_transfer_event = pallet_balances::Event::<Test>::Transfer {
+            let mut encoded_balance_transfer_event = pallet_balances::Event::<Test>::Transfer {
                 from: hex!("0909090909090909090909090909090909090909090909090909090909090909")
                     .into(), // variant A
                 to: hex!("0606060606060606060606060606060606060606060606060606060606060606").into(), // variant B (dest)
                 amount: 1, // variant A
             }
             .encode();
+
+            // Adding 4 since Balances Pallet = 4 in construct_runtime! enum
+            let mut encoded_event = vec![4];
+            encoded_event.append(&mut encoded_balance_transfer_event);
 
             println!(
                 "full side effects before confirmation: {:?}",
@@ -1191,7 +1598,7 @@ fn circuit_handles_transfer_and_swap() {
             assert_ok!(Circuit::confirm_side_effect(
                 origin_relayer_bob.clone(),
                 xtx_id.clone(),
-                valid_transfer_side_effect,
+                valid_transfer_side_effect_1,
                 confirmation_transfer,
                 None,
                 None,
@@ -1203,13 +1610,17 @@ fn circuit_handles_transfer_and_swap() {
             );
 
             // Confirmation start
-            let encoded_swap_transfer_event = orml_tokens::Event::<Test>::Transfer {
+            let mut encoded_swap_transfer_event = orml_tokens::Event::<Test>::Transfer {
                 currency_id: as_u32_le(&[0, 1, 2, 3]), // currency_id as u8 bytes [0,1,2,3] -> u32
                 from: BOB_RELAYER,                     // executor - Bob
                 to: hex!("0606060606060606060606060606060606060606060606060606060606060606").into(), // variant B (dest)
                 amount: 2u64, // amount - variant B
             }
             .encode();
+
+            // Adding 4 since Balances Pallet = 4 in construct_runtime! enum
+            let mut encoded_event = vec![4];
+            encoded_event.append(&mut encoded_swap_transfer_event);
 
             let confirmation_swap = ConfirmedSideEffect::<AccountId32, BlockNumber, BalanceOf> {
                 err: None,
@@ -1239,5 +1650,112 @@ fn circuit_handles_transfer_and_swap() {
                 None,
                 None,
             ));
+        });
+}
+
+#[test]
+fn circuit_cancels_xtx_after_timeout() {
+    let origin = Origin::signed(ALICE); // Only sudo access to register new gateways for now
+
+    let _origin_relayer_bob = Origin::signed(BOB_RELAYER); // Only sudo access to register new gateways for now
+
+    let transfer_protocol_box = ExtBuilder::get_transfer_protocol_box();
+    let _swap_protocol_box = ExtBuilder::get_swap_protocol_box();
+
+    let mut local_state = LocalState::new();
+    let valid_transfer_side_effect = produce_and_validate_side_effect(
+        vec![
+            (Type::Address(32), ArgVariant::A),
+            (Type::Address(32), ArgVariant::B),
+            (Type::Uint(64), ArgVariant::A),
+            (Type::Bytes(0), ArgVariant::A), // empty bytes instead of insurance
+        ],
+        &mut local_state,
+        transfer_protocol_box,
+    );
+
+    let side_effects = vec![valid_transfer_side_effect.clone()];
+    let fee = 1;
+    let sequential = false;
+
+    ExtBuilder::default()
+        .with_standard_side_effects()
+        .with_default_xdns_records()
+        .build()
+        .execute_with(|| {
+            let _ = Balances::deposit_creating(&ALICE, 10);
+
+            System::set_block_number(1);
+
+            assert_ok!(Circuit::on_extrinsic_trigger(
+                origin,
+                side_effects,
+                fee,
+                sequential,
+            ));
+
+            let _events = System::events();
+            // assert_eq!(events.len(), 8);
+
+            let xtx_id: sp_core::H256 =
+                hex!("c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6").into();
+
+            // The tiemout links that will be checked at on_initialize are there
+            assert_eq!(Circuit::get_active_timing_links(xtx_id), Some(100u64));
+
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id),
+                Some(XExecSignal {
+                    requester: AccountId32::new(hex!(
+                        "0101010101010101010101010101010101010101010101010101010101010101"
+                    )),
+                    timeouts_at: 100u64,
+                    delay_steps_at: None,
+                    status: CircuitStatus::Ready,
+                    total_reward: Some(fee),
+                    steps_cnt: (0, 1),
+                })
+            );
+
+            System::set_block_number(100);
+
+            <Circuit as frame_support::traits::OnInitialize<u64>>::on_initialize(100);
+
+            assert_eq!(
+                Circuit::get_x_exec_signals(xtx_id),
+                Some(XExecSignal {
+                    requester: AccountId32::new(hex!(
+                        "0101010101010101010101010101010101010101010101010101010101010101"
+                    )),
+                    timeouts_at: 100u64,
+                    delay_steps_at: None,
+                    status: CircuitStatus::RevertTimedOut,
+                    total_reward: Some(fee),
+                    steps_cnt: (0, 1),
+                })
+            );
+
+            assert_eq!(Circuit::get_active_timing_links(xtx_id), None);
+
+            // Emits event notifying about cancellation
+            let mut events = System::events();
+            // assert_eq!(events.len(), 9);
+            assert_eq!(
+                events.pop(),
+                Some(EventRecord {
+                    phase: Phase::Initialization,
+                    event: Event::Circuit(
+                        crate::Event::<Test>::XTransactionXtxRevertedAfterTimeOut(
+                            hex!(
+                                "c282160defd729da11b0cfcfed580278943723737b7017f56dbd32e695fc41e6"
+                            )
+                            .into()
+                        )
+                    ),
+                    topics: vec![]
+                }),
+            );
+
+            // Voids all associated side effects with Xtx by setting their confirmation to Err
         });
 }