+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## `v0.0.1-roco` 2022-04-20
+
+Rococo prerelease patch 1
+
+## `v0.0.0-roco` 2022-04-19
+
+Rococo prerelease
+
 ## `v1.0.0-alpha.0` 2020-10-21
 
 Initial milestone 2